requirements.txt
gitdb==4.0.10
GitPython==3.1.29
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
smmap==5.0.0
temp==2020.7.2
